+++ Libplanet/Crypto/DefaultCryptoBackend.cs
@@ -1,3 +1,4 @@
+using System;
 using System.IO;
 using System.Security.Cryptography;
 using Org.BouncyCastle.Asn1;
@@ -10,17 +11,24 @@ namespace Libplanet.Crypto
     public class DefaultCryptoBackend<T> : ICryptoBackend<T>
         where T : HashAlgorithm
     {
-        public byte[] Sign(HashDigest<T> messageHash, PrivateKey privateKey)
+        public byte[] Sign(HashDigest<T> messageHash, IPrivateKey privateKey)
         {
+            if (!(privateKey is PrivateKey pk))
+            {
+                throw new ArgumentException(
+                    $"Given {privateKey} is not {nameof(PrivateKey)}",
+                    nameof(privateKey));
+            }
+
             var h = new Sha256Digest();
             var kCalculator = new HMacDsaKCalculator(h);
             var signer = new ECDsaSigner(kCalculator);
-            signer.Init(true, privateKey.KeyParam);
+            signer.Init(true, pk.KeyParam);
             BigInteger[] rs = signer.GenerateSignature(messageHash.ToByteArray());
             var r = rs[0];
             var s = rs[1];
 
-            BigInteger otherS = privateKey.KeyParam.Parameters.N.Subtract(s);
+            BigInteger otherS = pk.KeyParam.Parameters.N.Subtract(s);
             if (s.CompareTo(otherS) == 1)
             {
                 s = otherS;
@@ -34,13 +42,17 @@ namespace Libplanet.Crypto
             return bos.ToArray();
         }
 
-        public bool Verify(
-            HashDigest<T> messageHash,
-            byte[] signature,
-            PublicKey publicKey)
+        public bool Verify(HashDigest<T> messageHash, byte[] signature, IPublicKey publicKey)
         {
             try
             {
+                if (!(publicKey is PublicKey pub))
+                {
+                    throw new ArgumentException(
+                        $"Given {publicKey} is not {nameof(PublicKey)}",
+                        nameof(publicKey));
+                }
+
                 Asn1Sequence asn1Sequence = (Asn1Sequence)Asn1Object.FromByteArray(signature);
 
                 var rs = new[]
@@ -49,7 +61,7 @@ namespace Libplanet.Crypto
                     ((DerInteger)asn1Sequence[1]).Value,
                 };
                 var verifier = new ECDsaSigner();
-                verifier.Init(false, publicKey.KeyParam);
+                verifier.Init(false, pub.KeyParam);
 
                 return verifier.VerifySignature(messageHash.ToByteArray(), rs[0], rs[1]);
             }